+++ db/install.php
@@ -236,7 +236,10 @@ function apply_constraints() {
             "FOREIGN KEY (user_id) REFERENCES {user} (id) ON DELETE CASCADE ON UPDATE CASCADE",
 
         "ALTER TABLE {charon_review_comment} ADD CONSTRAINT FK_charon_review_comment_submission_file " . 
-            "FOREIGN KEY (submission_file_id) REFERENCES {charon_submission_file} (id) ON DELETE CASCADE ON UPDATE CASCADE"
+            "FOREIGN KEY (submission_file_id) REFERENCES {charon_submission_file} (id) ON DELETE CASCADE ON UPDATE CASCADE",
+
+        "ALTER TABLE {charon_query_log_user} ADD CONSTRAINT FK_query_log_user_user " .
+            "FOREIGN KEY (user_id) REFERENCES {user} (id) ON DELETE CASCADE ON UPDATE CASCADE"
     );
 
     foreach ($cmds as $cmd) {